Refactor blacklist code to own feature module
- refactor commit moving blacklist stuff from `src/options/{containers,components}` to `src/options/blacklist`
- for general structure see: https://jaysoo.ca/2016/02/28/organizing-redux-application/#rule-1-organize-by-feature
- redux actions + reducers extracted from the `src/options/containers/settings` (should be main conflict with other UI work)
- everything exposed via `src/options/blacklist/index.js`
- update blacklist imports from external code
- blacklist state now stored separate in `state.blacklist` instead of `state.settings`
