--- src/options/components/blacklistRow/style.css
@@ -1,23 +1,0 @@
-.blacklistActions {
-    float: right;
-    position: relative;
-    right: -20px;
-
-    & a {
-        color: #999;
-        text-decoration: none;
-
-        &:hover {
-            color: #333;
-        }
-    }
-}
-
-.date {
-    font-size: .75em;
-}
-
-.icon {
-    font-size: 1.5em;
-    cursor: pointer;
-}
